requirements.txt
PyDictionary
google_trans_new
discord.py
discord.py[voice]
youtube-dl
google
dnspython
pymongo
lyricsgenius